+++ src/containers/Accounts/AccountTransactionsTable/actions.js
@@ -1,4 +1,4 @@
-import { getAccountTransactions } from '../../../rippled/routes/v1';
+import { getAccountTransactions } from '../../../rippled';
 import { analytics, ANALYTIC_TYPES } from '../../shared/utils';
 import * as actionTypes from './actionTypes';
