+++ src/containers/Ledger/actions.js
@@ -7,7 +7,7 @@ import {
   HASH_REGEX,
 } from '../shared/utils';
 import * as actionTypes from './actionTypes';
-import { getLedger } from '../../rippled/routes/v1';
+import { getLedger } from '../../rippled';
 
 export const loadLedger = identifier => dispatch => {
   if (!DECIMAL_REGEX.test(identifier) && !HASH_REGEX.test(identifier)) {
