+++ src/containers/Token/DEXPairs/index.js
@@ -11,7 +11,7 @@ import {
   analytics,
   ANALYTIC_TYPES,
 } from '../../shared/utils';
-import { getOffers } from '../../../rippled/routes/v1';
+import { getOffers } from '../../../rippled';
 import PairStats from './PairStats';
 
 // Hard Coded Pairs that we always check for
