+++ src/rippled/accountState.js
@@ -5,9 +5,9 @@
  */
 
 const addressCodec = require('ripple-address-codec');
-const rippled = require('../../lib/rippled');
-const log = require('../../lib/logger')({ name: 'account balances' });
-const utils = require('../../lib/utils');
+const rippled = require('./lib/rippled');
+const log = require('./lib/logger')({ name: 'account balances' });
+const utils = require('./lib/utils');
 
 const formatResults = (info, data) => {
   const balances = { XRP: Number(info.Balance) / 1000000 };
